requirement.txt
opencv-python
numpy
pyserial
adafruit-io
torch
fastapi[all]
uvicorn